+++ src/content-layers.test.js
@@ -1,16 +1,26 @@
-const { proj, sym, document, pres, root, isLayer } = require('./core');
+const {
+  proj,
+  sym,
+  document,
+  pres,
+  root,
+  isLayer,
+  seq,
+  str,
+  num,
+} = require('./core');
 
 describe('content layers', () => {
   const doc = () => {
     const d = document('base');
-    d.add(root, [1, 2, 3]);
+    d.add(root, seq([1, 2, 3]));
     d.add(1, sym('+'));
-    d.add(2, 1);
-    d.add(3, 2);
-    d.add(['layer1', 2], 11);
-    d.add(['layer1', 3], 21);
-    d.add(['layer1', 'layer1_1', 3], 211);
-    d.add(['layer2', 2], 12);
+    d.add(2, num(1));
+    d.add(3, num(2));
+    d.add(['layer1', 2], num(11));
+    d.add(['layer1', 3], num(21));
+    d.add(['layer1', 'layer1_1', 3], num(211));
+    d.add(['layer2', 2], num(12));
     return d;
   };
 
@@ -18,15 +28,15 @@ describe('content layers', () => {
     const d = doc();
     expect(d.value()).toMatchObject({
       1: sym('+'),
-      2: 1,
-      3: 2,
+      2: num(1),
+      3: num(2),
       layer1: {
-        2: 11,
-        3: 21,
-        layer1_1: { 3: 211 },
+        2: num(11),
+        3: num(21),
+        layer1_1: { 3: num(211) },
       },
       layer2: {
-        2: 12,
+        2: num(12),
       },
     });
     expect(isLayer(d.value())).toEqual(true);
@@ -40,8 +50,8 @@ describe('content layers', () => {
       const projection = proj(d);
       expect(projection).toMatchObject({
         1: sym('+'),
-        2: 1,
-        3: 2,
+        2: num(1),
+        3: num(2),
       });
       expect(projection.layer1).toEqual(undefined);
       expect(projection.layer2).toEqual(undefined);
@@ -52,8 +62,8 @@ describe('content layers', () => {
       const projection = proj(d, ['layer2']);
       expect(projection).toMatchObject({
         1: sym('+'),
-        2: 12,
-        3: 2,
+        2: num(12),
+        3: num(2),
       });
       expect(projection.layer1).toEqual(undefined);
       expect(projection.layer2).toEqual(undefined);
@@ -64,15 +74,15 @@ describe('content layers', () => {
       const projection = proj(d, ['layer1', 'layer2']);
       expect(projection).toMatchObject({
         1: sym('+'),
-        2: 12,
-        3: 21,
+        2: num(12),
+        3: num(21),
       });
 
       const projection2 = proj(d, ['layer2', 'layer1']);
       expect(projection2).toMatchObject({
         1: sym('+'),
-        2: 11,
-        3: 21,
+        2: num(11),
+        3: num(21),
       });
     });
 
@@ -81,46 +91,46 @@ describe('content layers', () => {
       const projection = proj(d, [['layer1', ['layer1_1']]]);
       expect(projection).toMatchObject({
         1: sym('+'),
-        2: 11,
-        3: 211,
+        2: num(11),
+        3: num(211),
       });
     });
 
     it('root replacement', () => {
       const d = document('test');
-      d.add(root, 'a');
-      d.add(['replacement', root], 'b');
+      d.add(root, str('a'));
+      d.add(['replacement', root], str('b'));
       expect(proj(d, ['replacement'])).toMatchObject({
-        [root]: 'b',
+        [root]: str('b'),
       });
     });
 
     it('metalayers', () => {
       const d = document('test');
-      d.add(root, [1, 2]);
-      d.add(1, 'a');
-      d.add(['descr', 1], 'small a');
-      d.add(['ts', 1], 1588321340608);
-      d.add(2, 'b');
-      d.add(['descr', 2], 'small b');
-      d.add(['ts', 2], 1588321366606);
-      d.add(['alt', 2], 'B');
-      d.add(['alt', 'descr', 2], 'big b');
+      d.add(root, seq([1, 2]));
+      d.add(1, str('a'));
+      d.add(['descr', 1], str('small a'));
+      d.add(['ts', 1], num(1588321340608));
+      d.add(2, str('b'));
+      d.add(['descr', 2], str('small b'));
+      d.add(['ts', 2], num(1588321366606));
+      d.add(['alt', 2], str('B'));
+      d.add(['alt', 'descr', 2], str('big b'));
       const result = proj(d, ['alt'], ['descr', 'ts']);
       expect(result).toMatchObject({
-        [root]: [1, 2],
-        1: 'a',
-        2: 'B',
+        [root]: seq([1, 2]),
+        1: str('a'),
+        2: str('B'),
         descr: {
-          1: 'small a',
-          2: 'big b',
+          1: str('small a'),
+          2: str('big b'),
         },
         ts: {
-          1: 1588321340608,
+          1: num(1588321340608),
           // todo: ponder about the desired value in this case
           // Do we really want to keep the metavalue of a layer below,
           // if the value is not set?
-          2: 1588321366606,
+          2: num(1588321366606),
         },
       });
     });
@@ -129,6 +139,6 @@ describe('content layers', () => {
   it('pres', () => {
     const d = doc();
     const projection = proj(d, [['layer1', ['layer1_1']]]);
-    expect(pres(projection)).toMatchObject([sym('+'), 11, 211]);
+    expect(pres(projection)).toMatchObject([sym('+'), num(11), num(211)]);
   });
 });